refactor: 인생지도 페이지 성능 개선 (#183)
* chore: 안쓰는 폰트 제거

* chore: 블러 이미지 -> css로 변경

* refactor: 지도 영역에 dynamic import로 변경

* refactor: 지도 페이지 하단 영역 분리 및 dynamic import

* refactor: 빈 카드 이미지에 priority 추가

* refactor: 지도 페이지 런타임 edge로 변경

* chore: import문 수정

* chore: use client 제거

* refactor: 번들 사이즈 개선 (#184)

* refactor: lottie 번들 사이즈 축소

* chore: sentry 제거
