+++ .pnp.cjs
@@ -30,8 +30,6 @@ const RAW_RUNTIME_STATE =
           ["@radix-ui/react-icons", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:1.3.0"],\
           ["@radix-ui/react-portal", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:1.0.4"],\
           ["@radix-ui/react-tabs", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:1.0.4"],\
-          ["@sentry/nextjs", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"],\
           ["@storybook/addon-essentials", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:7.5.3"],\
           ["@storybook/addon-interactions", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:7.5.3"],\
           ["@storybook/addon-links", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:7.5.3"],\
@@ -7086,281 +7084,11 @@ const RAW_RUNTIME_STATE =
         "linkType": "HARD"\
       }]\
     ]],\
-    ["@rollup/plugin-commonjs", [\
-      ["npm:24.0.0", {\
-        "packageLocation": "./.yarn/cache/@rollup-plugin-commonjs-npm-24.0.0-b18d79acac-10c0.zip/node_modules/@rollup/plugin-commonjs/",\
-        "packageDependencies": [\
-          ["@rollup/plugin-commonjs", "npm:24.0.0"]\
-        ],\
-        "linkType": "SOFT"\
-      }],\
-      ["virtual:f85a24cd7249cbf55993a720e319ff07eec708b84af0539161ec0749b373c97ed32ae133a24dd3d6c3ae19b98d63d2acc657d2aadb999fdda2a9c2705e78151d#npm:24.0.0", {\
-        "packageLocation": "./.yarn/__virtual__/@rollup-plugin-commonjs-virtual-ff91a8a65f/0/cache/@rollup-plugin-commonjs-npm-24.0.0-b18d79acac-10c0.zip/node_modules/@rollup/plugin-commonjs/",\
-        "packageDependencies": [\
-          ["@rollup/plugin-commonjs", "virtual:f85a24cd7249cbf55993a720e319ff07eec708b84af0539161ec0749b373c97ed32ae133a24dd3d6c3ae19b98d63d2acc657d2aadb999fdda2a9c2705e78151d#npm:24.0.0"],\
-          ["@rollup/pluginutils", "virtual:ff91a8a65fddf63f5ad1da1e7430edce2244efab0cd67092fac1a11078121174be44c28b151ccc8f2c5baf3b2b6f1950deae41cdf4947220675888e357e56b19#npm:5.1.0"],\
-          ["@types/rollup", null],\
-          ["commondir", "npm:1.0.1"],\
-          ["estree-walker", "npm:2.0.2"],\
-          ["glob", "npm:8.1.0"],\
-          ["is-reference", "npm:1.2.1"],\
-          ["magic-string", "npm:0.27.0"],\
-          ["rollup", "npm:2.78.0"]\
-        ],\
-        "packagePeers": [\
-          "@types/rollup",\
-          "rollup"\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@rollup/pluginutils", [\
-      ["npm:5.1.0", {\
-        "packageLocation": "./.yarn/cache/@rollup-pluginutils-npm-5.1.0-6939820ef8-10c0.zip/node_modules/@rollup/pluginutils/",\
-        "packageDependencies": [\
-          ["@rollup/pluginutils", "npm:5.1.0"]\
-        ],\
-        "linkType": "SOFT"\
-      }],\
-      ["virtual:ff91a8a65fddf63f5ad1da1e7430edce2244efab0cd67092fac1a11078121174be44c28b151ccc8f2c5baf3b2b6f1950deae41cdf4947220675888e357e56b19#npm:5.1.0", {\
-        "packageLocation": "./.yarn/__virtual__/@rollup-pluginutils-virtual-87b72af0d7/0/cache/@rollup-pluginutils-npm-5.1.0-6939820ef8-10c0.zip/node_modules/@rollup/pluginutils/",\
-        "packageDependencies": [\
-          ["@rollup/pluginutils", "virtual:ff91a8a65fddf63f5ad1da1e7430edce2244efab0cd67092fac1a11078121174be44c28b151ccc8f2c5baf3b2b6f1950deae41cdf4947220675888e357e56b19#npm:5.1.0"],\
-          ["@types/estree", "npm:1.0.5"],\
-          ["@types/rollup", null],\
-          ["estree-walker", "npm:2.0.2"],\
-          ["picomatch", "npm:2.3.1"],\
-          ["rollup", "npm:2.78.0"]\
-        ],\
-        "packagePeers": [\
-          "@types/rollup",\
-          "rollup"\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
     ["@rushstack/eslint-patch", [\
       ["npm:1.5.1", {\
         "packageLocation": "./.yarn/cache/@rushstack-eslint-patch-npm-1.5.1-caed8f5804-10c0.zip/node_modules/@rushstack/eslint-patch/",\
         "packageDependencies": [\
           ["@rushstack/eslint-patch", "npm:1.5.1"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry-internal/feedback", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-internal-feedback-npm-7.91.0-c26a8e27b8-10c0.zip/node_modules/@sentry-internal/feedback/",\
-        "packageDependencies": [\
-          ["@sentry-internal/feedback", "npm:7.91.0"],\
-          ["@sentry/core", "npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry-internal/tracing", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-internal-tracing-npm-7.91.0-3f7019e90f-10c0.zip/node_modules/@sentry-internal/tracing/",\
-        "packageDependencies": [\
-          ["@sentry-internal/tracing", "npm:7.91.0"],\
-          ["@sentry/core", "npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/browser", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-browser-npm-7.91.0-cecb31f0ff-10c0.zip/node_modules/@sentry/browser/",\
-        "packageDependencies": [\
-          ["@sentry/browser", "npm:7.91.0"],\
-          ["@sentry-internal/feedback", "npm:7.91.0"],\
-          ["@sentry-internal/tracing", "npm:7.91.0"],\
-          ["@sentry/core", "npm:7.91.0"],\
-          ["@sentry/replay", "npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/cli", [\
-      ["npm:1.77.1", {\
-        "packageLocation": "./.yarn/unplugged/@sentry-cli-npm-1.77.1-ffa1e2e09c/node_modules/@sentry/cli/",\
-        "packageDependencies": [\
-          ["@sentry/cli", "npm:1.77.1"],\
-          ["https-proxy-agent", "npm:5.0.1"],\
-          ["mkdirp", "npm:0.5.6"],\
-          ["node-fetch", "virtual:017b14e33dc24496ace239ade2b5640ed536b1269040d3967889e91041f773a863c46aedbc2262a41a584ccc9a00fc7c284e30859c940aa03a3ba62080987605#npm:2.7.0"],\
-          ["progress", "npm:2.0.3"],\
-          ["proxy-from-env", "npm:1.1.0"],\
-          ["which", "npm:2.0.2"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/core", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-core-npm-7.91.0-67ba791446-10c0.zip/node_modules/@sentry/core/",\
-        "packageDependencies": [\
-          ["@sentry/core", "npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/integrations", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-integrations-npm-7.91.0-0670312a7c-10c0.zip/node_modules/@sentry/integrations/",\
-        "packageDependencies": [\
-          ["@sentry/integrations", "npm:7.91.0"],\
-          ["@sentry/core", "npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"],\
-          ["localforage", "npm:1.10.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/nextjs", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-nextjs-npm-7.91.0-5203cd0f56-10c0.zip/node_modules/@sentry/nextjs/",\
-        "packageDependencies": [\
-          ["@sentry/nextjs", "npm:7.91.0"]\
-        ],\
-        "linkType": "SOFT"\
-      }],\
-      ["virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:7.91.0", {\
-        "packageLocation": "./.yarn/__virtual__/@sentry-nextjs-virtual-f85a24cd72/0/cache/@sentry-nextjs-npm-7.91.0-5203cd0f56-10c0.zip/node_modules/@sentry/nextjs/",\
-        "packageDependencies": [\
-          ["@sentry/nextjs", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:7.91.0"],\
-          ["@rollup/plugin-commonjs", "virtual:f85a24cd7249cbf55993a720e319ff07eec708b84af0539161ec0749b373c97ed32ae133a24dd3d6c3ae19b98d63d2acc657d2aadb999fdda2a9c2705e78151d#npm:24.0.0"],\
-          ["@sentry/core", "npm:7.91.0"],\
-          ["@sentry/integrations", "npm:7.91.0"],\
-          ["@sentry/node", "npm:7.91.0"],\
-          ["@sentry/react", "virtual:f85a24cd7249cbf55993a720e319ff07eec708b84af0539161ec0749b373c97ed32ae133a24dd3d6c3ae19b98d63d2acc657d2aadb999fdda2a9c2705e78151d#npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"],\
-          ["@sentry/vercel-edge", "npm:7.91.0"],\
-          ["@sentry/webpack-plugin", "npm:1.21.0"],\
-          ["@types/next", null],\
-          ["@types/react", "npm:18.2.37"],\
-          ["@types/webpack", null],\
-          ["chalk", "npm:3.0.0"],\
-          ["next", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:14.0.2"],\
-          ["react", "npm:18.2.0"],\
-          ["resolve", "patch:resolve@npm%3A1.22.8#optional!builtin<compat/resolve>::version=1.22.8&hash=c3c19d"],\
-          ["rollup", "npm:2.78.0"],\
-          ["stacktrace-parser", "npm:0.1.10"],\
-          ["webpack", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:5.89.0"]\
-        ],\
-        "packagePeers": [\
-          "@types/next",\
-          "@types/react",\
-          "@types/webpack",\
-          "next",\
-          "react",\
-          "webpack"\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/node", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-node-npm-7.91.0-7c7ccbc13d-10c0.zip/node_modules/@sentry/node/",\
-        "packageDependencies": [\
-          ["@sentry/node", "npm:7.91.0"],\
-          ["@sentry-internal/tracing", "npm:7.91.0"],\
-          ["@sentry/core", "npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"],\
-          ["https-proxy-agent", "npm:5.0.1"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/react", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-react-npm-7.91.0-e4c9e245ff-10c0.zip/node_modules/@sentry/react/",\
-        "packageDependencies": [\
-          ["@sentry/react", "npm:7.91.0"]\
-        ],\
-        "linkType": "SOFT"\
-      }],\
-      ["virtual:f85a24cd7249cbf55993a720e319ff07eec708b84af0539161ec0749b373c97ed32ae133a24dd3d6c3ae19b98d63d2acc657d2aadb999fdda2a9c2705e78151d#npm:7.91.0", {\
-        "packageLocation": "./.yarn/__virtual__/@sentry-react-virtual-4dda987820/0/cache/@sentry-react-npm-7.91.0-e4c9e245ff-10c0.zip/node_modules/@sentry/react/",\
-        "packageDependencies": [\
-          ["@sentry/react", "virtual:f85a24cd7249cbf55993a720e319ff07eec708b84af0539161ec0749b373c97ed32ae133a24dd3d6c3ae19b98d63d2acc657d2aadb999fdda2a9c2705e78151d#npm:7.91.0"],\
-          ["@sentry/browser", "npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"],\
-          ["@types/react", "npm:18.2.37"],\
-          ["hoist-non-react-statics", "npm:3.3.2"],\
-          ["react", "npm:18.2.0"]\
-        ],\
-        "packagePeers": [\
-          "@types/react",\
-          "react"\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/replay", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-replay-npm-7.91.0-057a5ece6d-10c0.zip/node_modules/@sentry/replay/",\
-        "packageDependencies": [\
-          ["@sentry/replay", "npm:7.91.0"],\
-          ["@sentry-internal/tracing", "npm:7.91.0"],\
-          ["@sentry/core", "npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/types", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-types-npm-7.91.0-cff635972d-10c0.zip/node_modules/@sentry/types/",\
-        "packageDependencies": [\
-          ["@sentry/types", "npm:7.91.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/utils", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-utils-npm-7.91.0-ee32f0b121-10c0.zip/node_modules/@sentry/utils/",\
-        "packageDependencies": [\
-          ["@sentry/utils", "npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/vercel-edge", [\
-      ["npm:7.91.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-vercel-edge-npm-7.91.0-9ecf1e2451-10c0.zip/node_modules/@sentry/vercel-edge/",\
-        "packageDependencies": [\
-          ["@sentry/vercel-edge", "npm:7.91.0"],\
-          ["@sentry-internal/tracing", "npm:7.91.0"],\
-          ["@sentry/core", "npm:7.91.0"],\
-          ["@sentry/types", "npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["@sentry/webpack-plugin", [\
-      ["npm:1.21.0", {\
-        "packageLocation": "./.yarn/cache/@sentry-webpack-plugin-npm-1.21.0-0d5f07e5b7-10c0.zip/node_modules/@sentry/webpack-plugin/",\
-        "packageDependencies": [\
-          ["@sentry/webpack-plugin", "npm:1.21.0"],\
-          ["@sentry/cli", "npm:1.77.1"],\
-          ["webpack-sources", "npm:3.2.3"]\
         ],\
         "linkType": "HARD"\
       }]\
@@ -10677,14 +10405,6 @@ const RAW_RUNTIME_STATE =
         ],\
         "linkType": "HARD"\
       }],\
-      ["npm:6.0.2", {\
-        "packageLocation": "./.yarn/cache/agent-base-npm-6.0.2-428f325a93-10c0.zip/node_modules/agent-base/",\
-        "packageDependencies": [\
-          ["agent-base", "npm:6.0.2"],\
-          ["debug", "virtual:2f4f5d4be0c4271f81a756602d102e02c292a5b30d8b7f5dd19d5368d75df8bb2f1bf516e3f835c2b5501d1663eb437d4f00d6d057bcaee770b46e87267a0fa3#npm:4.3.4"]\
-        ],\
-        "linkType": "HARD"\
-      }],\
       ["npm:7.1.0", {\
         "packageLocation": "./.yarn/cache/agent-base-npm-7.1.0-4b12ba5111-10c0.zip/node_modules/agent-base/",\
         "packageDependencies": [\
@@ -10803,8 +10523,6 @@ const RAW_RUNTIME_STATE =
           ["@radix-ui/react-icons", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:1.3.0"],\
           ["@radix-ui/react-portal", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:1.0.4"],\
           ["@radix-ui/react-tabs", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:1.0.4"],\
-          ["@sentry/nextjs", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:7.91.0"],\
-          ["@sentry/utils", "npm:7.91.0"],\
           ["@storybook/addon-essentials", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:7.5.3"],\
           ["@storybook/addon-interactions", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:7.5.3"],\
           ["@storybook/addon-links", "virtual:74868250102727df59d0c08f1ddb6630e6cd48640d8ad9d1ed80f220eee5d95c37bdc80e67cbbd686a520d72c8276bea707c0a77572e4e760d1051e7254c23b3#npm:7.5.3"],\
@@ -12184,15 +11902,6 @@ const RAW_RUNTIME_STATE =
           ["ansi-styles", "npm:3.2.1"],\
           ["escape-string-regexp", "npm:1.0.5"],\
           ["supports-color", "npm:5.5.0"]\
-        ],\
-        "linkType": "HARD"\
-      }],\
-      ["npm:3.0.0", {\
-        "packageLocation": "./.yarn/cache/chalk-npm-3.0.0-e813208025-10c0.zip/node_modules/chalk/",\
-        "packageDependencies": [\
-          ["chalk", "npm:3.0.0"],\
-          ["ansi-styles", "npm:4.3.0"],\
-          ["supports-color", "npm:7.2.0"]\
         ],\
         "linkType": "HARD"\
       }],\
@@ -14551,15 +14260,6 @@ const RAW_RUNTIME_STATE =
         "linkType": "HARD"\
       }]\
     ]],\
-    ["estree-walker", [\
-      ["npm:2.0.2", {\
-        "packageLocation": "./.yarn/cache/estree-walker-npm-2.0.2-dfab42f65c-10c0.zip/node_modules/estree-walker/",\
-        "packageDependencies": [\
-          ["estree-walker", "npm:2.0.2"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
     ["esutils", [\
       ["npm:2.0.3", {\
         "packageLocation": "./.yarn/cache/esutils-npm-2.0.3-f865beafd5-10c0.zip/node_modules/esutils/",\
@@ -15487,18 +15187,6 @@ const RAW_RUNTIME_STATE =
           ["path-is-absolute", "npm:1.0.1"]\
         ],\
         "linkType": "HARD"\
-      }],\
-      ["npm:8.1.0", {\
-        "packageLocation": "./.yarn/cache/glob-npm-8.1.0-65f64af8b1-10c0.zip/node_modules/glob/",\
-        "packageDependencies": [\
-          ["glob", "npm:8.1.0"],\
-          ["fs.realpath", "npm:1.0.0"],\
-          ["inflight", "npm:1.0.6"],\
-          ["inherits", "npm:2.0.4"],\
-          ["minimatch", "npm:5.1.6"],\
-          ["once", "npm:1.4.0"]\
-        ],\
-        "linkType": "HARD"\
       }]\
     ]],\
     ["glob-parent", [\
@@ -15899,15 +15587,6 @@ const RAW_RUNTIME_STATE =
         ],\
         "linkType": "HARD"\
       }],\
-      ["npm:5.0.1", {\
-        "packageLocation": "./.yarn/cache/https-proxy-agent-npm-5.0.1-42d65f358e-10c0.zip/node_modules/https-proxy-agent/",\
-        "packageDependencies": [\
-          ["https-proxy-agent", "npm:5.0.1"],\
-          ["agent-base", "npm:6.0.2"],\
-          ["debug", "virtual:2f4f5d4be0c4271f81a756602d102e02c292a5b30d8b7f5dd19d5368d75df8bb2f1bf516e3f835c2b5501d1663eb437d4f00d6d057bcaee770b46e87267a0fa3#npm:4.3.4"]\
-        ],\
-        "linkType": "HARD"\
-      }],\
       ["npm:7.0.2", {\
         "packageLocation": "./.yarn/cache/https-proxy-agent-npm-7.0.2-83ea6a5d42-10c0.zip/node_modules/https-proxy-agent/",\
         "packageDependencies": [\
@@ -16014,15 +15693,6 @@ const RAW_RUNTIME_STATE =
         "packageDependencies": [\
           ["image-size", "npm:1.0.2"],\
           ["queue", "npm:6.0.2"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["immediate", [\
-      ["npm:3.0.6", {\
-        "packageLocation": "./.yarn/cache/immediate-npm-3.0.6-c27588a2d3-10c0.zip/node_modules/immediate/",\
-        "packageDependencies": [\
-          ["immediate", "npm:3.0.6"]\
         ],\
         "linkType": "HARD"\
       }]\
@@ -16443,16 +16113,6 @@ const RAW_RUNTIME_STATE =
         "packageLocation": "./.yarn/cache/is-plain-object-npm-5.0.0-285b70faa3-10c0.zip/node_modules/is-plain-object/",\
         "packageDependencies": [\
           ["is-plain-object", "npm:5.0.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["is-reference", [\
-      ["npm:1.2.1", {\
-        "packageLocation": "./.yarn/cache/is-reference-npm-1.2.1-87ca1743c8-10c0.zip/node_modules/is-reference/",\
-        "packageDependencies": [\
-          ["is-reference", "npm:1.2.1"],\
-          ["@types/estree", "npm:1.0.5"]\
         ],\
         "linkType": "HARD"\
       }]\
@@ -17576,16 +17236,6 @@ const RAW_RUNTIME_STATE =
         "linkType": "HARD"\
       }]\
     ]],\
-    ["lie", [\
-      ["npm:3.1.1", {\
-        "packageLocation": "./.yarn/cache/lie-npm-3.1.1-91350720d9-10c0.zip/node_modules/lie/",\
-        "packageDependencies": [\
-          ["lie", "npm:3.1.1"],\
-          ["immediate", "npm:3.0.6"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
     ["lilconfig", [\
       ["npm:2.1.0", {\
         "packageLocation": "./.yarn/cache/lilconfig-npm-2.1.0-a179261924-10c0.zip/node_modules/lilconfig/",\
@@ -17684,16 +17334,6 @@ const RAW_RUNTIME_STATE =
         "packageLocation": "./.yarn/cache/loader-utils-npm-3.2.1-76ae2fd253-10c0.zip/node_modules/loader-utils/",\
         "packageDependencies": [\
           ["loader-utils", "npm:3.2.1"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["localforage", [\
-      ["npm:1.10.0", {\
-        "packageLocation": "./.yarn/cache/localforage-npm-1.10.0-cf9ea9a436-10c0.zip/node_modules/localforage/",\
-        "packageDependencies": [\
-          ["localforage", "npm:1.10.0"],\
-          ["lie", "npm:3.1.1"]\
         ],\
         "linkType": "HARD"\
       }]\
@@ -17854,16 +17494,6 @@ const RAW_RUNTIME_STATE =
         "packageLocation": "./.yarn/cache/lz-string-npm-1.5.0-3860794e30-10c0.zip/node_modules/lz-string/",\
         "packageDependencies": [\
           ["lz-string", "npm:1.5.0"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["magic-string", [\
-      ["npm:0.27.0", {\
-        "packageLocation": "./.yarn/cache/magic-string-npm-0.27.0-a60a83c0b4-10c0.zip/node_modules/magic-string/",\
-        "packageDependencies": [\
-          ["magic-string", "npm:0.27.0"],\
-          ["@jridgewell/sourcemap-codec", "npm:1.4.15"]\
         ],\
         "linkType": "HARD"\
       }]\
@@ -20892,16 +20522,6 @@ const RAW_RUNTIME_STATE =
         "linkType": "HARD"\
       }]\
     ]],\
-    ["rollup", [\
-      ["npm:2.78.0", {\
-        "packageLocation": "./.yarn/cache/rollup-npm-2.78.0-09284f4c78-10c0.zip/node_modules/rollup/",\
-        "packageDependencies": [\
-          ["rollup", "npm:2.78.0"],\
-          ["fsevents", "patch:fsevents@npm%3A2.3.3#optional!builtin<compat/fsevents>::version=2.3.3&hash=df0bf1"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
     ["run-applescript", [\
       ["npm:5.0.0", {\
         "packageLocation": "./.yarn/cache/run-applescript-npm-5.0.0-ea4b8840dd-10c0.zip/node_modules/run-applescript/",\
@@ -21451,16 +21071,6 @@ const RAW_RUNTIME_STATE =
         "packageLocation": "./.yarn/cache/stackframe-npm-1.3.4-bf4b7cc8fd-10c0.zip/node_modules/stackframe/",\
         "packageDependencies": [\
           ["stackframe", "npm:1.3.4"]\
-        ],\
-        "linkType": "HARD"\
-      }]\
-    ]],\
-    ["stacktrace-parser", [\
-      ["npm:0.1.10", {\
-        "packageLocation": "./.yarn/cache/stacktrace-parser-npm-0.1.10-36f3e571bd-10c0.zip/node_modules/stacktrace-parser/",\
-        "packageDependencies": [\
-          ["stacktrace-parser", "npm:0.1.10"],\
-          ["type-fest", "npm:0.7.1"]\
         ],\
         "linkType": "HARD"\
       }]\
@@ -22556,13 +22166,6 @@ const RAW_RUNTIME_STATE =
         "packageLocation": "./.yarn/cache/type-fest-npm-0.6.0-76b229965b-10c0.zip/node_modules/type-fest/",\
         "packageDependencies": [\
           ["type-fest", "npm:0.6.0"]\
-        ],\
-        "linkType": "HARD"\
-      }],\
-      ["npm:0.7.1", {\
-        "packageLocation": "./.yarn/cache/type-fest-npm-0.7.1-7b37912923-10c0.zip/node_modules/type-fest/",\
-        "packageDependencies": [\
-          ["type-fest", "npm:0.7.1"]\
         ],\
         "linkType": "HARD"\
       }],\
