+++ package.json
@@ -18,8 +18,6 @@
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-portal": "^1.0.4",
     "@radix-ui/react-tabs": "^1.0.4",
-    "@sentry/nextjs": "^7.91.0",
-    "@sentry/utils": "^7.91.0",
     "@svgr/webpack": "^8.1.0",
     "@tanstack/react-query": "^5.8.4",
     "@toss/use-overlay": "^1.3.8",
