+++ src/app/fonts.ts
@@ -3,18 +3,6 @@ import localFont from 'next/font/local';
 export const pretendard = localFont({
   src: [
     {
-      path: '../assets/fonts/Pretendard/Pretendard-Thin.woff2',
-      weight: '100',
-    },
-    {
-      path: '../assets/fonts/Pretendard/Pretendard-ExtraLight.woff2',
-      weight: '200',
-    },
-    {
-      path: '../assets/fonts/Pretendard/Pretendard-Light.woff2',
-      weight: '300',
-    },
-    {
       path: '../assets/fonts/Pretendard/Pretendard-Regular.woff2',
       weight: '400',
     },
@@ -29,14 +17,6 @@ export const pretendard = localFont({
     {
       path: '../assets/fonts/Pretendard/Pretendard-Bold.woff2',
       weight: '700',
-    },
-    {
-      path: '../assets/fonts/Pretendard/Pretendard-ExtraBold.woff2',
-      weight: '800',
-    },
-    {
-      path: '../assets/fonts/Pretendard/Pretendard-Black.woff2',
-      weight: '900',
     },
   ],
   variable: '--font-pretendard',
