+++ src/app/home/[username]/layout.tsx
@@ -1,8 +1,6 @@
 import type { PropsWithChildren } from 'react';
 import type { Metadata } from 'next';
-import Image from 'next/image';
 
-import PurpleBlurImage from '@/assets/images/purple-blur.png';
 import { getMetadata } from '@/utils/getMetadata';
 
 import type { HomeRouteParams } from './page';
@@ -14,9 +12,7 @@ export const generateMetadata = async ({ params: { username } }: HomeRouteParams
 const HomeLayout = ({ children }: PropsWithChildren) => {
   return (
     <main className="relative flex justify-center bg-gradient1">
-      <div className="absolute right-0">
-        <Image src={PurpleBlurImage} width={390} height={125} alt="purple_blur" priority />
-      </div>
+      <div className="absolute right-0 w-full h-[125px] bg-gradientPurpleBlur" />
       <div className="w-full h-[100dvh]">
         <div className="w-full h-[100dvh] flex flex-col items-center justify-between pb-xs">{children}</div>
       </div>
@@ -25,3 +21,5 @@ const HomeLayout = ({ children }: PropsWithChildren) => {
 };
 
 export default HomeLayout;
+
+export const runtime = 'edge';
