+++ src/features/home/components/mapSwiper/MapSwiper.tsx
@@ -17,7 +17,7 @@ interface MapSwiperProps {
   currentPage: number | null;
 }
 
-export const MapSwiper = ({ currentPage, children }: PropsWithChildren<MapSwiperProps>) => {
+const MapSwiper = ({ currentPage, children }: PropsWithChildren<MapSwiperProps>) => {
   return (
     currentPage !== null && (
       <Swiper {...settings} initialSlide={currentPage} className="h-full">
@@ -27,3 +27,5 @@ export const MapSwiper = ({ currentPage, children }: PropsWithChildren<MapSwiper
     )
   );
 };
+
+export default MapSwiper;
