--- src/features/home/components/mapSwiper/index.ts
@@ -1,1 +1,0 @@
-export { MapSwiper } from './MapSwiper';
