+++ src/features/home/components/shareBottomSheet/index.ts
@@ -1,1 +1,2 @@
+export { ImageDownloadBottomSheet } from './ImageDownloadBottomSheet';
 export { ShareBottomSheet } from './ShareBottomSheet';
